Fix #400: move autoloader to src/Google
This makes system-wide installation of the library more viable.
For backward compatibility, an autoload.php is preserved in the
old location, which prints a deprecation warning and then includes
the autoloader from the new location.
