+++ src/Google/Auth/AppIdentity.php
@@ -22,7 +22,7 @@
  */
 use google\appengine\api\app_identity\AppIdentityService;
 
-require_once realpath(dirname(__FILE__) . '/../../../autoload.php');
+require_once realpath(dirname(__FILE__) . '/../autoload.php');
 
 /**
  * Authentication via the Google App Engine App Identity service.
