+++ src/Google/Auth/AssertionCredentials.php
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-require_once realpath(dirname(__FILE__) . '/../../../autoload.php');
+require_once realpath(dirname(__FILE__) . '/../autoload.php');
 
 /**
  * Credentials object used for OAuth 2.0 Signed JWT assertion grants.
