+++ tests/OAuthHelper.php
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-require_once dirname(__FILE__) . '/../autoload.php';
+require_once dirname(__FILE__) . '/../src/Google/autoload.php';
 
 $client = new Google_Client();
 $client->setScopes(
