+++ tests/bootstrap.php
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-require_once dirname(__FILE__) . '/../autoload.php';
+require_once dirname(__FILE__) . '/../src/Google/autoload.php';
 require_once dirname(__FILE__) . '/BaseTest.php';
 
 date_default_timezone_set('UTC');
